move oc adm commands to staging
