+++ pkg/oc/cli/admin/admin.go
@@ -13,6 +13,11 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	ktemplates "k8s.io/kubernetes/pkg/kubectl/util/templates"
 
+	"github.com/openshift/oc/pkg/cli/admin/createerrortemplate"
+	"github.com/openshift/oc/pkg/cli/admin/createlogintemplate"
+	"github.com/openshift/oc/pkg/cli/admin/createproviderselectiontemplate"
+	"github.com/openshift/oc/pkg/cli/admin/node"
+	"github.com/openshift/oc/pkg/cli/admin/upgrade"
 	"github.com/openshift/oc/pkg/cli/options"
 	cmdutil "github.com/openshift/oc/pkg/helpers/cmd"
 	"github.com/openshift/origin/pkg/cmd/server/admin"
@@ -20,9 +25,6 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/admin/buildchain"
 	"github.com/openshift/origin/pkg/oc/cli/admin/cert"
 	"github.com/openshift/origin/pkg/oc/cli/admin/createbootstrapprojecttemplate"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createerrortemplate"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createlogintemplate"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createproviderselectiontemplate"
 	"github.com/openshift/origin/pkg/oc/cli/admin/groups"
 	"github.com/openshift/origin/pkg/oc/cli/admin/migrate"
 	migrateetcd "github.com/openshift/origin/pkg/oc/cli/admin/migrate/etcd"
@@ -32,13 +34,11 @@ import (
 	migratetemplateinstances "github.com/openshift/origin/pkg/oc/cli/admin/migrate/templateinstances"
 	"github.com/openshift/origin/pkg/oc/cli/admin/mustgather"
 	"github.com/openshift/origin/pkg/oc/cli/admin/network"
-	"github.com/openshift/origin/pkg/oc/cli/admin/node"
 	"github.com/openshift/origin/pkg/oc/cli/admin/policy"
 	"github.com/openshift/origin/pkg/oc/cli/admin/project"
 	"github.com/openshift/origin/pkg/oc/cli/admin/prune"
 	"github.com/openshift/origin/pkg/oc/cli/admin/release"
 	"github.com/openshift/origin/pkg/oc/cli/admin/top"
-	"github.com/openshift/origin/pkg/oc/cli/admin/upgrade"
 	"github.com/openshift/origin/pkg/oc/cli/admin/verifyimagesignature"
 	"github.com/openshift/origin/pkg/oc/cli/kubectlwrappers"
 )
