+++ staging/src/github.com/openshift/oc/pkg/cli/admin/createlogintemplate/create_login_template.go
